refactor(bot): isolate execution policy
